fix(provider/gce): Ensure referenced instances exist during L4 delete.
