+++ clouddriver-google/src/main/groovy/com/netflix/spinnaker/clouddriver/google/deploy/GCEUtil.groovy
@@ -28,6 +28,7 @@ import com.google.api.services.compute.Compute
 import com.google.api.services.compute.model.*
 import com.netflix.spinnaker.cats.cache.Cache
 import com.netflix.spinnaker.cats.cache.RelationshipCacheFilter
+import com.netflix.spinnaker.clouddriver.consul.provider.ConsulProviderUtils
 import com.netflix.spinnaker.clouddriver.data.task.Task
 import com.netflix.spinnaker.clouddriver.google.GoogleConfiguration
 import com.netflix.spinnaker.clouddriver.google.GoogleExecutorTraits
@@ -40,6 +41,7 @@ import com.netflix.spinnaker.clouddriver.google.deploy.exception.GoogleOperation
 import com.netflix.spinnaker.clouddriver.google.deploy.exception.GoogleResourceNotFoundException
 import com.netflix.spinnaker.clouddriver.google.model.*
 import com.netflix.spinnaker.clouddriver.google.model.callbacks.Utils
+import com.netflix.spinnaker.clouddriver.google.model.health.GoogleInstanceHealth
 import com.netflix.spinnaker.clouddriver.google.model.loadbalancing.*
 import com.netflix.spinnaker.clouddriver.google.provider.view.GoogleClusterProvider
 import com.netflix.spinnaker.clouddriver.google.provider.view.GoogleLoadBalancerProvider
@@ -1777,5 +1779,66 @@ class GCEUtil {
       healthChecks.addAll(healthCheckList.getItems() ?: [])
     }
     return healthChecks
+  }
+
+  static List<GoogleInstance> fetchInstances(GoogleExecutorTraits agent, GoogleNamedAccountCredentials credentials) {
+    List<GoogleInstance> instances = new ArrayList<GoogleInstance>()
+    String pageToken = null
+
+    while (true) {
+      InstanceAggregatedList instanceAggregatedList = agent.timeExecute(
+        credentials.compute.instances().aggregatedList(credentials.project).setPageToken(pageToken),
+        "compute.instances.aggregatedList",
+        agent.TAG_SCOPE, agent.SCOPE_GLOBAL)
+
+      instances += transformInstances(instanceAggregatedList, credentials)
+      pageToken = instanceAggregatedList.getNextPageToken()
+
+      if (!pageToken) {
+        break
+      }
+    }
+
+    return instances
+  }
+
+  private static List<GoogleInstance> transformInstances(InstanceAggregatedList instanceAggregatedList, GoogleNamedAccountCredentials credentials) throws IOException {
+    List<GoogleInstance> instances = []
+
+    instanceAggregatedList?.items?.each { String zone, InstancesScopedList instancesScopedList ->
+      def localZoneName = Utils.getLocalName(zone)
+      instancesScopedList?.instances?.each { Instance instance ->
+        def consulNode = credentials.consulConfig?.enabled ?
+          ConsulProviderUtils.getHealths(credentials.consulConfig, instance.getName())
+          : null
+        long instanceTimestamp = instance.creationTimestamp ?
+          Utils.getTimeFromTimestamp(instance.creationTimestamp) :
+          Long.MAX_VALUE
+        String instanceName = Utils.getLocalName(instance.name)
+        def googleInstance = new GoogleInstance(
+          name: instanceName,
+          gceId: instance.id,
+          instanceType: Utils.getLocalName(instance.machineType),
+          cpuPlatform: instance.cpuPlatform,
+          launchTime: instanceTimestamp,
+          zone: localZoneName,
+          region: credentials.regionFromZone(localZoneName),
+          networkInterfaces: instance.networkInterfaces,
+          networkName: Utils.decorateXpnResourceIdIfNeeded(credentials.project, instance.networkInterfaces?.getAt(0)?.network),
+          metadata: instance.metadata,
+          disks: instance.disks,
+          serviceAccounts: instance.serviceAccounts,
+          selfLink: instance.selfLink,
+          tags: instance.tags,
+          labels: instance.labels,
+          consulNode: consulNode,
+          instanceHealth: new GoogleInstanceHealth(
+            status: GoogleInstanceHealth.Status.valueOf(instance.getStatus())
+          ))
+        instances << googleInstance
+      }
+    }
+
+    return instances
   }
 }
